make separate module for u256 and address in interface
